move queue fiels
