+++ codegen.yml
@@ -15,7 +15,7 @@ generates:
     plugins:
       - add: '/* THIS IS A GENERATED FILE */'
       - add: '/* /Maybe<\(Maybe<(.*)>\)\[\]>/ */'
-      - add: '/* /Maybe<$1[]>/ */'
+      - add: '/* Maybe<$1[]> */'
       - typescript-common
       - typescript-client
   graphql.schema.json:
