+++ src/graphql.d.ts
@@ -89,16 +89,6 @@ declare module '*/PlayerAnime.graphql' {
   export default defaultDocument
 }
 
-declare module '*/Queue.graphql' {
-  import { DocumentNode } from 'graphql'
-  const defaultDocument: DocumentNode
-  const Queue: DocumentNode
-
-  export { Queue }
-
-  export default defaultDocument
-}
-
 declare module '*/RewatchMutation.graphql' {
   import { DocumentNode } from 'graphql'
   const defaultDocument: DocumentNode
@@ -185,6 +175,16 @@ declare module '*/list.graphql' {
   const ListView: DocumentNode
 
   export { ListView }
+
+  export default defaultDocument
+}
+
+declare module '*/queue.graphql' {
+  import { DocumentNode } from 'graphql'
+  const defaultDocument: DocumentNode
+  const Queue: DocumentNode
+
+  export { Queue }
 
   export default defaultDocument
 }
