+++ src/views/queue/components/source-select.vue
@@ -150,7 +150,7 @@ export default class SourceSelect extends Vue {
 </script>
 
 <style scoped lang="scss">
-@import '../../colors';
+@import '../../../colors';
 
 .source-select {
   position: relative;
